move netmanage to package
