+++ html/collectors/cisco_asa_collectors.html
@@ -31,7 +31,7 @@
 import ansible_runner
 import pandas as pd
 
-from helpers import helpers as hp
+from netmanage.helpers import helpers as hp
 
 
 def get_interface_ips(username: str,
