+++ html/collectors/cisco_ios_collectors.html
@@ -31,7 +31,7 @@
 import ansible_runner
 import pandas as pd
 
-from helpers import helpers as hp
+from netmanage.helpers import helpers as hp
 
 
 def gather_facts(username: str,
