+++ html/collectors/cisco_nxos_collectors.html
@@ -34,7 +34,7 @@ import pandas as pd
 import re
 import sqlite3 as sl
 
-from helpers import helpers as hp
+from netmanage.helpers import helpers as hp
 
 
 def nxos_diff_running_config(username: str,
