+++ html/collectors/palo_alto_collectors.html
@@ -31,7 +31,7 @@
 import ansible_runner
 import json
 import pandas as pd
-from helpers import helpers as hp
+from netmanage.helpers import helpers as hp
 from typing import Dict
