+++ html/validators.html
@@ -35,7 +35,7 @@ Runs data collectors and stores them in a sqlite database.
 
 import pandas as pd
 import sqlite3 as sl
-from helpers import helpers as hp
+from netmanage.helpers import helpers as hp
 
 
 def f5_node_availability(db_path: str, table: str) -&gt; pd.DataFrame:
