+++ netmanage/__init__.py
@@ -1,0 +1,7 @@
+from . import run_collectors # noqa
+from netmanage.setup import select_hostgroups # noqa
+
+__all__ = [
+    'run_collectors',
+    'select_hostgroups'
+]
