+++ netmanage/collectors/cisco_asa_collectors.py
@@ -3,7 +3,7 @@
 import ansible_runner
 import pandas as pd
 
-from helpers import helpers as hp
+from netmanage.helpers import helpers as hp
 
 
 def get_interface_ips(username: str,
