+++ netmanage/collectors/palo_alto_collectors.py
@@ -3,7 +3,7 @@
 import ansible_runner
 import json
 import pandas as pd
-from helpers import helpers as hp
+from netmanage.helpers import helpers as hp
 from typing import Dict
