+++ netmanage/helpers/helpers.py
