+++ test/test_f5_collectors.py
@@ -7,7 +7,7 @@ import sys
 nm_path = os.environ.get('NM_PATH')
 os.chdir(f'{nm_path}/test')
 sys.path.append('..')
-from collectors import f5_collectors as collectors  # noqa
+from netmanage.collectors import f5_collectors as collectors  # noqa
 
 
 def test_get_arp_table(username,
